+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Move ci to tests
 * Move CSS to assets
 * Move files to assets and includes
+* No longer require standalone free plugin activation
 * Readme installation
 * RESOLVE Non-SSL protocol used for scripts and styles
 * Revise headers